Fix bug that caused properties to be read before we knew where to read them from
The Launcher was initiating SimpleSynchronize in it's constructor, before parsing
command line arguments. However, SimpleSynchronize's constructor tries to read the
configuration, which may lead to an error (no config file found) or reading the 
wrong config file (if one is in the classpath).

Reworked the code to parse the most basic command line options first (including
config location), and then start a SimpleSynchronize and pass along command line 
options.

This didn't appear in the 1.1 branch.